feat(jobserver): Pass cxt config to manager start
This feature adds back something that was changed in the refactoring
around starting up in cluster mode, namely, it gives you access to the
entire context configuration.

The reason for this change is that it allows for more flexibility in
scheduling your driver.

For example, in yarn or mesos mode, we might not need a very large SJS
instance in terms of memory, but would like a large driver for some
contexts. Currently, there is no way to get information at runtime about
the context to create. By passing this configuration, the manager_start
script can inspect the config, extract values, and pass the appropriate
options (such as `--driver-memory` and `--driver-cores`) to
spark-submit.

This is implemented such that it comes before the optional proxy-user
argument. This fixes the default manager_start.sh command but could be a
slightly breaking change to any customized manager_start.sh scripts that
make use of proxy-user.

The other option is to put it after the optional proxy-user command, but
that might be a bit more clunky. Feedback appreciated!