feat(jobserver) Don't remove hostname in cluster
Currently, when you try and run in cluster mode and have multiple
contexts it blows up all the existing contexts.

The reason for this is that when a jobManager initially joins, it
connects to the master and they work fine. However, when you try to join
a second jobManager, all the jobManagers and SJS actors try to connect
to one another, but since the JobManager is broadcasting a localhost
address, when it tries to create the full mesh in akka, it can't and
akka starts marking the nodes unreachable.

The correct answer here is to interpolate something like `HOSTNAME` and
not use any hardcoded values